chore: Remove footer and move links to status bar MAASENG-2058
