--- src/app/base/components/Footer/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./Footer";
